Move entirety of prologue_epilogue logic to abi module.
